Version 5 - switched default log level to "EMERGENCY" to reduce initial performance load.
Hopefully fixed profileactions misbehavior - it was only showing the logs link after install, but corrected itself with an QRR.

Added a check for the SugarGator to not log entries sent to the sugarcrm log - these can be very frequent and verbose, so it's good to be able to turn them off.

Minor fix for the PruneLogsRecords job to reference the table_name property instead of hard-coding the table name.

Updates for manifest and README.