Added new dashlet for managing SugarGator config values.
Added a SugarGatorApi to support changing config values from the UI.

Changed the channel and log_level fields to be enums instead of text fields.

Added functions to the sg_LogsAggregator module to return options for the existing channels and log levels to support changing those fields to enums instead of text fields.

Added a visibility grid to the log_level field so that log levels specific to the sugarcrm.log won't be shown for monolog channel loggers and vice-versa.

Hacked a custom enum field type for the sg_LogsAggregator bean to support the visibility grid. The problem is that the channel field is the trigger for the visibility grid and we have no way of knowing in advance what channels will exist.
So this custom field will decide if the channel is 'sugarcrm' or not, and if not, return the 'not_sugarcrm' options.

Added some defensive coding to SugarGator.php to prevent infinite loop logging.
Changed the SugarGator to always set the log_level value to lower case, to match expected values from config.

Changed the index for all SugarGator config arrays to be 'SugarGator' instead of an integer because the integer values seem to bork the configurator.